+++ src/main/java/net/spell_engine/SpellEngineMod.java
@@ -8,8 +8,10 @@ import net.minecraft.advancement.criterion.Criteria;
 import net.minecraft.client.resource.language.I18n;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.spell_engine.api.effect.StatusEffectClassification;
 import net.spell_engine.api.spell.ExternalSpellSchools;
+import net.spell_engine.api.spell.event.SpellEvents;
 import net.spell_engine.api.spell.registry.SpellRegistry;
 import net.spell_engine.config.ServerConfig;
 import net.spell_engine.config.ServerConfigWrapper;
@@ -18,6 +20,7 @@ import net.spell_engine.internals.SpellTriggers;
 import net.spell_engine.internals.container.SpellAssignments;
 import net.spell_engine.internals.container.SpellContainerSource;
 import net.spell_engine.internals.criteria.EnchantmentSpecificCriteria;
+import net.spell_engine.internals.criteria.SpellCastCriteria;
 import net.spell_engine.internals.delivery.SpellStashHelper;
 import net.spell_engine.network.ServerNetwork;
 import net.spell_engine.fx.SpellEngineParticles;
@@ -43,6 +46,10 @@ public class SpellEngineMod {
         SpellEngineParticles.register();
 
         Criteria.register(EnchantmentSpecificCriteria.ID.toString(), EnchantmentSpecificCriteria.INSTANCE);
+        Criteria.register(SpellCastCriteria.ID.toString(), SpellCastCriteria.INSTANCE);
+        SpellEvents.SPELL_CAST.register(args -> {
+            SpellCastCriteria.INSTANCE.trigger((ServerPlayerEntity) args.caster(), args.spell());
+        });
 
         ExternalSpellSchools.init();
         RPGSeriesCore.init();